Refactor lib into utils
