+++ pkg/utils/other.go
@@ -1,0 +1,15 @@
+package utils
+
+import (
+	"fmt"
+	"time"
+)
+
+// FormatDate formats time into
+// DD-MM-YYYY format
+func FormatDate(t time.Time) string {
+	y, m, d := t.Date()
+	// keep last 2 digits in year
+	y %= 100
+	return fmt.Sprintf("%02d/%02d/%d", d, m, y)
+}
